use  build tag to switch modes, instead of flag
